rv32: implement D, and the NaN-boxing it makes F responsible for
Closes #9. FLEN is 64, the arithmetic is Berkeley SoftFloat like F, and
both backends pass the whole official suite: **378/378, interpreter and
--jit alike**, where it was 274/274 before -- the 104 D tests are new.
riscv-tests stays 77/77.

Your instruction was to back D with the host FPU in the JIT where one
exists. I did not, and the reason is in CLAUDE.md: that arrangement was
tried for *single* precision and reverted. A host FPU asked to be a
RISC-V FPU failed 23 of the 78 F tests -- NaN propagation, subnormals,
and which of fflags an operation may raise -- so everything that
rounds, classifies or reports a flag goes to rv_hart_fp from both
backends. D inherits that, which is why the JIT needed no arithmetic
work at all. The door to speed is the one that entry describes: bring
operations back one at a time, each measured against the suite.

**NaN-boxing is the actual work, and it is a change to F rather than an
addition beside it.** A single value in a 64-bit register must carry
all-ones in its upper half or read as the canonical NaN, so every F
operation is affected in both directions -- and the failures do not
point at the instruction that got it wrong. The next *reader* sees the
NaN, so F-fsub.s fails because something else wrote badly. Two live
cases survived converting the obvious ones: the fused multiply-adds
wrote h->f[rd] raw, and FMV.X.W has to move the low half **raw**
rather than through the unboxing accessor, because moving bits is the
whole instruction.

It also cost the JIT its two exceptions. FLW/FSW and FMV.X.W/FMV.W.X
were lowered natively through EMU_IR_FGET/FPUT, which move 32 bits and
cannot box: 93 of 378 tests. They go to the helper now.

FLD/FSD are eight bytes on a 32-bit bus, so two accesses, with both
loads completing before either half of the register is written -- the
same shape as G4MH's LD.DW. Zcd comes with them, because C@2.0 includes
the compressed double load/stores when D is present exactly as it
includes Zcf's when F is, and UDB rejects the config without it. Their
immediate scales by eight where C.FLW's scales by four, which offset 0
cannot tell apart, so the decode tests use offsets that can.

Declared to UDB *and* Sail, both, and the run said why that matters
twice over. Missing MUTABLE_MISA_D failed DUT prep while the runner
still reported "All 274 tests passed" -- the previous ELFs, rebuilt
from nothing. And ACT selects suites by directory name, so D had to be
named in run-arch-test.sh before a single D test was built. A green
result from a suite that did not grow is the null result this file
keeps recording.

Cost: +9,308 bytes of F746 text, 147,364 to 156,672. -DRV32_EXT_D=OFF
gives back the old core exactly, misa and the configure banner
included.

+++ include/rv32/rv_hart.h
@@ -57,11 +57,21 @@ typedef struct rv_hart {
 
 #if RV_EXT_F
     /*
-     * Single-precision register file. 32 bits per register because D is not
-     * implemented; with D these would be 64-bit and F values NaN-boxed into
-     * the low half.
+     * The floating-point register file, 64 bits wide whether or not D is
+     * built in.
+     *
+     * With D, FLEN is 64 and a single-precision value is **NaN-boxed**:
+     * its upper 32 bits are all ones, and a register holding anything
+     * else reads as the canonical NaN when used as a float. That is not
+     * an addition beside the F operations -- it is a change to every one
+     * of them, which is why the accessors in rv_fpu.c exist and why
+     * nothing outside that file touches this array.
+     *
+     * Without D the upper half is unused. Keeping the array one width
+     * costs 128 bytes on the target and saves a second set of code
+     * paths, which is the trade this project makes elsewhere too.
      */
-    uint32_t f[32];
+    uint64_t f[32];
     uint32_t fcsr;               /* frm in [7:5], fflags in [4:0] */
 #endif
